requirements.txt
requests>=2.32.0
python-dotenv>=1.0.0
pillow>=10.0.0
anthropic>=0.40.0
pytz>=2024.1
